Consolidate web form tests into one suite
